discrepancy: fix warnings, etc; JIRA:RW-1246
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@92628 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/discrepancy_report/asndisc.cpp
@@ -565,7 +565,7 @@ int main(int argc, const char* argv[])
     if (argc==2 && argv && argv[1] && strchr(argv[1], ' '))
     {
         NStr::Split(argv[1], " ", split_args);
-        argc = 1 + split_args.size();
+        argc = 1 + (int)split_args.size();
         new_argv.reserve(argc);
         new_argv.push_back(argv[0]);
         for (auto& s: split_args)
